+++ enterprise/api/v1/__init__.py
@@ -1,0 +1,3 @@
+"""
+API endpoint for enterprise app.
+"""