+++ enterprise/models.py
@@ -236,7 +236,9 @@ class EnterpriseCustomerUser(TimeStampedModel):
         user_id (:class:`django.db.models.IntegerField`): user identifier
     """
 
-    enterprise_customer = models.ForeignKey(EnterpriseCustomer, blank=False, null=False)
+    enterprise_customer = models.ForeignKey(
+        EnterpriseCustomer, blank=False, null=False, related_name='enterprise_customer_users'
+    )
     user_id = models.PositiveIntegerField(null=False, blank=False)
 
     objects = EnterpriseCustomerUserManager()
@@ -514,7 +516,7 @@ class UserDataSharingConsentAudit(TimeStampedModel):
         (DISABLED, 'Disabled'),
     )
 
-    user = models.ForeignKey(EnterpriseCustomerUser)
+    user = models.ForeignKey(EnterpriseCustomerUser, related_name='data_sharing_consent')
 
     state = models.CharField(
         max_length=8,
@@ -646,7 +648,7 @@ class EnterpriseCourseEnrollment(TimeStampedModel):
             # If it is indeterminate...
 
             # Check for an account-wide value and use that.
-            consent_state = self.enterprise_customer_user.userdatasharingconsentaudit_set.first()
+            consent_state = self.enterprise_customer_user.data_sharing_consent.first()
             if consent_state is not None:
                 return consent_state.enabled
             else: