+++ enterprise/urls.py
@@ -4,7 +4,7 @@ URLs for enterprise.
 """
 from __future__ import absolute_import, unicode_literals
 
-from django.conf.urls import url
+from django.conf.urls import include, url
 
 from enterprise.views import GrantDataSharingPermissions
 
@@ -14,4 +14,9 @@ urlpatterns = [
         GrantDataSharingPermissions.as_view(),
         name='grant_data_sharing_permissions'
     ),
+    url(
+        r'^enterprise/api/',
+        include('enterprise.api.urls'),
+        name='enterprise_api'
+    )
 ]